Use int value for register id.
As different version of protocol may have different
register and lower/upper case, it is hard and time-
consuming to decode it at runtime.

With this change, we can use int value for register id
and the protocol will be decoded at startup once.

This would accelerate the runtime decoding and make it
easier to read.

Signed-off-by: maybetaken <296496296@qq.com>