Hotfix for unload device fail
Unload device fail may requires user reboot homeassistant
which is a not friendly way for users.

This fix only unload platforms that a certain device loaded.

This issue is quite interesting, as switch sensor is also not
used, but it didn't reports any error while unloading devices,
my guess for this root cause is said 'Time' sensor is not
that compatible with current HA system. It upstream fixed this
is, this change may not be required.

Signed-off-by: maybetaken <296496296@qq.com>